refactor(upgradeIconV1): split out transform icon
