+++ packages/components/codemods/upgradeIconV1/upgradeIconV1.spec.ts
@@ -91,6 +91,44 @@ describe("upgradeIconV1()", () => {
     })
   })
 
+  it("transforms old Icon", () => {
+    const inputAst = parseJsx(`
+      import { FlagOnIcon } from "@kaizen/components"
+      export const TestComponent = () => <FlagOnIcon />
+    `)
+    const outputAst = parseJsx(`
+      import { Icon } from "@kaizen/components/future"
+      export const TestComponent = () => <Icon name="flag" isFilled />
+    `)
+    expect(
+      transformIcons(
+        inputAst,
+        new Map([
+          ["@kaizen/components", new Map([["FlagOnIcon", "FlagOnIcon"]])],
+        ])
+      )
+    ).toEqual(printAst(outputAst))
+  })
+
+  it("transforms aliased old Icon", () => {
+    const inputAst = parseJsx(`
+      import { HamburgerIcon as IconAlias } from "@kaizen/components"
+      export const TestComponent = () => <IconAlias />
+    `)
+    const outputAst = parseJsx(`
+      import { Icon } from "@kaizen/components/future"
+      export const TestComponent = () => <Icon name="menu" />
+    `)
+    expect(
+      transformIcons(
+        inputAst,
+        new Map([
+          ["@kaizen/components", new Map([["IconAlias", "HamburgerIcon"]])],
+        ])
+      )
+    ).toEqual(printAst(outputAst))
+  })
+
   describe("import statements", () => {
     it("does not update import of components which are not Icons", () => {
       const inputAst = parseJsx(`
@@ -105,15 +143,7 @@ describe("upgradeIconV1()", () => {
       expect(
         transformIcons(
           inputAst,
-          new Map([
-            [
-              "@kaizen/components",
-              new Map([
-                ["AddIcon", "AddIcon"],
-                ["Card", "Card"],
-              ]),
-            ],
-          ])
+          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
         )
       ).toEqual(printAst(outputAst))
     })
@@ -154,252 +184,6 @@ describe("upgradeIconV1()", () => {
             ],
             ["somewhere-else", new Map([["HamHam", "HamburgerIcon"]])],
           ])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-  })
-
-  it("renames component and adds equivalent props", () => {
-    const inputAst = parseJsx(`
-      import { AddIcon, FlagOffIcon, FlagOffWhiteIcon, FlagOnIcon, HamburgerIcon as IconAlias, MeatballsIcon } from "@kaizen/components"
-      export const TestComponent = () => (
-        <>
-          <AddIcon />
-          <FlagOffIcon />
-          <FlagOffWhiteIcon />
-          <FlagOnIcon />
-          <IconAlias />
-          <MeatballsIcon />
-        </>
-      )
-    `)
-    const outputAst = parseJsx(`
-      import { Icon } from "@kaizen/components/future"
-      export const TestComponent = () => (
-        <>
-          <Icon name="add" />
-          <Icon name="flag" />
-          <Icon name="flag" />
-          <Icon name="flag" isFilled />
-          <Icon name="menu" />
-          <Icon name="more_horiz" />
-        </>
-      )
-    `)
-    expect(
-      transformIcons(
-        inputAst,
-        new Map([
-          [
-            "@kaizen/components",
-            new Map([
-              ["AddIcon", "AddIcon"],
-              ["FlagOffIcon", "FlagOffIcon"],
-              ["FlagOffWhiteIcon", "FlagOffWhiteIcon"],
-              ["FlagOnIcon", "FlagOnIcon"],
-              ["IconAlias", "HamburgerIcon"],
-              ["MeatballsIcon", "MeatballsIcon"],
-            ]),
-          ],
-        ])
-      )
-    ).toEqual(printAst(outputAst))
-  })
-
-  it("leaves icons which do not have a new equivalent", () => {
-    const inputAst = parseJsx(`
-      import { SomeInvalidIcon } from "@kaizen/components"
-      export const TestComponent = () => (
-        <SomeInvalidIcon />
-      )
-    `)
-    const outputAst = parseJsx(`
-      import { SomeInvalidIcon } from "@kaizen/components"
-      export const TestComponent = () => (
-        <SomeInvalidIcon />
-      )
-    `)
-    expect(
-      transformIcons(
-        inputAst,
-        new Map([
-          [
-            "@kaizen/components",
-            new Map([["SomeInvalidationIcon", "SomeInvalidationIcon"]]),
-          ],
-        ])
-      )
-    ).toEqual(printAst(outputAst))
-  })
-
-  describe("transform existing props", () => {
-    it("replaces role and aria-label with equivalent prop", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => (
-          <>
-            <AddIcon role="presentation" />
-            <AddIcon role="img" aria-label="Add something" />
-            <AddIcon role="img" aria-label={alt} />
-          </>
-        )
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => (
-          <>
-            <Icon name="add" isPresentational />
-            <Icon name="add" alt="Add something" />
-            <Icon name="add" alt={alt} />
-          </>
-        )
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-
-    it("replaces classNameOverride with className prop", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => <AddIcon role="presentation" classNameOverride="mt-16" />
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => <Icon name="add" isPresentational className="mt-16" />
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-
-    it("leaves inheritSize - this should throw a TS error for the consumer", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => <AddIcon role="presentation" inheritSize />
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => <Icon name="add" isPresentational
-        // @todo: Apply the correct --icon-size (eg. in Tailwind: className="[--icon-size:48]")
-        inheritSize />
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-
-    it("removes aria-hidden", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => <AddIcon role="presentation" aria-hidden={true} />
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => <Icon name="add" isPresentational />
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-
-    it("removes fontSize", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => <AddIcon role="presentation" fontSize={20} />
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => <Icon name="add" isPresentational />
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-
-    it("removes viewBox", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => <AddIcon role="presentation" viewBox="0 0 24 24" />
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => <Icon name="add" isPresentational />
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-        )
-      ).toEqual(printAst(outputAst))
-    })
-
-    describe("color prop to style", () => {
-      it("transforms a string value", () => {
-        const inputAst = parseJsx(`
-          export const TestComponent = () => (
-            <>
-              <AddIcon color="grey" />
-              <AddIcon color="#0168b3" />
-            </>
-          )
-        `)
-        const outputAst = parseJsx(`
-          export const TestComponent = () => (
-            <>
-              <Icon name="add" style={{ color: "grey" }} />
-              <Icon name="add" style={{ color: "#0168b3" }} />
-            </>
-          )
-        `)
-        expect(
-          transformIcons(
-            inputAst,
-            new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-          )
-        ).toEqual(printAst(outputAst))
-      })
-
-      it("transforms a variable", () => {
-        const inputAst = parseJsx(`
-          export const TestComponent = () => (
-            <>
-              <AddIcon color={c.gray500} />
-              <AddIcon color={variableGrey} />
-            </>
-          )
-        `)
-        const outputAst = parseJsx(`
-          export const TestComponent = () => (
-            <>
-              <Icon name="add" style={{ color: c.gray500 }} />
-              <Icon name="add" style={{ color: variableGrey }} />
-            </>
-          )
-        `)
-        expect(
-          transformIcons(
-            inputAst,
-            new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
-          )
-        ).toEqual(printAst(outputAst))
-      })
-    })
-
-    it("transforms height and width to style", () => {
-      const inputAst = parseJsx(`
-        export const TestComponent = () => <AddIcon height={24} width={24} />
-      `)
-      const outputAst = parseJsx(`
-        export const TestComponent = () => <Icon name="add" style={{ height: 24, width: 24 }} />
-      `)
-      expect(
-        transformIcons(
-          inputAst,
-          new Map([["@kaizen/components", new Map([["AddIcon", "AddIcon"]])]])
         )
       ).toEqual(printAst(outputAst))
     })
